[FEATURE] (login): rework how this app handles login and authentication
There are two scenarios to consider: Navigation from Mobile device and form Desktop device.
From a desktop, we always navigate to a deep page like "/Start" or "/WE" (= not login page). Users are always authenticated by SAP NetWeaver.
However, from a mobile device, we always navigate to "/Anmeldung" (= login page) while being authenticated by SAP NetWeaver using an default dummy user. This is why we force the user to enter his username. On successful login, we set the user model. To prevent navigation to a deeper page without an user model, all pages are based on PageBaseConroller that attaches a onBeforeNavigation event. Each time this even triggers, we check if the user model is set and valid. If not, we navigate to login page.
To prevent desktop devices to be forced to navigatio to login page, we will read the IllumLoginName and fire a validation request to MII.